BUG: Fix bugs with reject caching
