+++ mne_bids_pipeline/scripts/preprocessing/_04a_run_ica.py
@@ -34,6 +34,7 @@ from config import (make_epochs, gen_log_kwargs, failsafe_run, _script_path,
                     annotations_to_events, _update_for_splits)
 from config import parallel_func
 
+from ..._utils import _get_reject, _write_json
 
 logger = logging.getLogger('mne-bids-pipeline')
 
@@ -266,6 +267,8 @@ def run_ica(*, cfg, subject, session, in_files):
         suffix='ica', extension='.fif')
     out_files['components'] = bids_basename.copy().update(
         processing='ica', suffix='components', extension='.tsv')
+    out_files['reject'] = bids_basename.copy().update(
+        processing='ica', suffix='reject', extension='.json')
     out_files['report'] = bids_basename.copy().update(
         processing='ica+components', suffix='report', extension='.html')
     del bids_basename
@@ -382,15 +385,23 @@ def run_ica(*, cfg, subject, session, in_files):
         epochs.set_eeg_reference(cfg.eeg_reference, projection=projection)
 
     # Reject epochs based on peak-to-peak rejection thresholds
-    msg = f'Using PTP rejection thresholds: {cfg.ica_reject}'
+    ica_reject = _get_reject(
+        subject=subject,
+        session=session,
+        epochs=epochs,
+        reject=cfg.ica_reject,
+        ch_types=cfg.ch_types,
+        decim=cfg.ica_decim)
+
+    msg = f'Using PTP rejection thresholds: {ica_reject}'
     logger.info(**gen_log_kwargs(message=msg, subject=subject,
                                  session=session))
 
-    epochs.drop_bad(reject=cfg.ica_reject)
+    epochs.drop_bad(reject=ica_reject)
     if epochs_eog is not None:
-        epochs_eog.drop_bad(reject=cfg.ica_reject)
+        epochs_eog.drop_bad(reject=ica_reject)
     if epochs_ecg is not None:
-        epochs_ecg.drop_bad(reject=cfg.ica_reject)
+        epochs_ecg.drop_bad(reject=ica_reject)
 
     # Now actually perform ICA.
     msg = 'Calculating ICA solution.'
@@ -501,6 +512,8 @@ def run_ica(*, cfg, subject, session, in_files):
     report.save(
         out_files['report'], overwrite=True, open_browser=cfg.interactive)
 
+    _write_json(out_files['reject'], ica_reject)
+
     assert len(in_files) == 0, in_files.keys()
     return out_files
 
@@ -525,7 +538,7 @@ def get_config(
         ica_n_components=config.ica_n_components,
         ica_max_iterations=config.ica_max_iterations,
         ica_decim=config.ica_decim,
-        ica_reject=config.get_ica_reject(),
+        ica_reject=config.ica_reject,
         ica_eog_threshold=config.ica_eog_threshold,
         ica_ctps_ecg_threshold=config.ica_ctps_ecg_threshold,
         random_state=config.random_state,
