+++ mne_bids_pipeline/scripts/preprocessing/_05a_apply_ica.py
@@ -27,6 +27,7 @@ import config
 from config import gen_log_kwargs, failsafe_run, _update_for_splits
 from config import parallel_func, _script_path
 
+from ..._utils import _read_json
 
 logger = logging.getLogger('mne-bids-pipeline')
 
@@ -53,6 +54,8 @@ def get_input_fnames_apply_ica(**kwargs):
         processing='ica', suffix='components', extension='.tsv')
     in_files['epochs'] = bids_basename.copy().update(
         suffix='epo', extension='.fif')
+    in_files['reject'] = bids_basename.copy().update(
+        processing='ica', suffix='reject', extension='.json')
     return in_files
 
 
@@ -92,7 +95,8 @@ def apply_ica(*, cfg, subject, session, in_files):
                                  session=session))
 
     epochs = mne.read_epochs(in_files.pop('epochs'), preload=True)
-    epochs.drop_bad(cfg.ica_reject)
+    ica_reject = _read_json(in_files.pop('reject'))
+    epochs.drop_bad(ica_reject)
 
     # Now actually reject the components.
     msg = f'Rejecting ICs: {", ".join([str(ic) for ic in ica.exclude])}'
@@ -145,7 +149,7 @@ def get_config(
         deriv_root=config.get_deriv_root(),
         interactive=config.interactive,
         baseline=config.baseline,
-        ica_reject=config.get_ica_reject(),
+        ica_reject=config.ica_reject,
         _epochs_split_size=config._epochs_split_size,
     )
     return cfg
