+++ mne_bids_pipeline/scripts/report/_01_make_reports.py
@@ -27,6 +27,7 @@ from config import (
     get_noise_cov_bids_path, _update_for_splits, _restrict_analyze_channels,
 )
 
+from ..._utils import _read_json
 
 logger = logging.getLogger('mne-bids-pipeline')
 
@@ -624,8 +625,13 @@ def run_report_preprocessing(
             )
         )
         epochs = mne.read_epochs(fname_epo_not_clean)
-        epochs.drop_bad(cfg.ica_reject)
         ica = mne.preprocessing.read_ica(fname_ica)
+        fname_reject = bids_path.copy().update(
+            processing='ica', suffix='reject', extension='.json')
+        reject = _read_json(fname_reject)
+        # TODO: Ref is set during ICA epochs fitting, we should ensure we do
+        # it here, too
+        epochs.drop_bad(reject)
 
         if ica.exclude:
             report.add_ica(
@@ -1605,7 +1611,6 @@ def get_config(
         conditions=config.conditions,
         all_contrasts=config.get_all_contrasts(),
         decoding_contrasts=config.get_decoding_contrasts(),
-        ica_reject=config.get_ica_reject(),
         time_frequency_conditions=config.time_frequency_conditions,
         decode=config.decode,
         decoding_metric=config.decoding_metric,
